requirements.txt
amqp~=2.2.2
arrow~=0.12.1
Babel~=2.5.3
backpack~=0.1
billiard~=3.5.0.3
blinker~=1.4
cachy~=0.2.0
celery~=4.1.0
certifi~=2018.4.16
chardet~=3.0.4
cleo~=0.6.5
DateTime~=4.2
dateutils~=0.6.6
eventlet~=0.22.1
Faker~=0.7.18
flower~=0.9.2
gevent>=1.2.2
greenlet~=0.4.13
idna~=2.6
inflection~=0.3.1
kombu~=4.1.0
lazy-object-proxy~=1.3.1
orator~=0.9.7
orator-cache~=0.1.1
pastel~=0.1.0
pendulum~=1.5.1
psutil~=5.4.5
pyaml~=16.12.2
Pygments~=2.2.0
pylev~=1.3.0
pymitter~=0.2.3
PyMySQL~=0.8.0
python-dateutil~=2.7.2
python-memcached~=1.59
pytz~=2018.4
pytzdata~=2018.4
PyYAML~=3.12
redis~=2.10.6
requests~=2.18.4
simplejson~=3.14.0
six~=1.11.0
stringutils~=0.3.0
thrift~=0.11.0
tornado~=5.0.2
tzlocal~=1.5.1
urllib3~=1.22
var-dump~=1.2
vine~=1.1.4
wrapt~=1.10.11
zope.interface~=4.5.0
